docs/website/tools/requirements.txt
beautifulsoup4
plotly
ruamel.yaml
pandas
pyzotero
